requirements.txt
bcrypt==3.1.3
Flask==0.12.1
Jinja2==2.9.6
MySQL-python==1.2.5
PyYAML==3.12
Werkzeug==0.12.1